requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
pymupdf
docx2txt
